feat: refactored main screen to support multiple chains
