+++ VultisigApp/VultisigApp/Features/Defi/DefiChain/Interactor/Bond/BondInteractor.swift
@@ -1,0 +1,11 @@
+//
+//  BondInteractor.swift
+//  VultisigApp
+//
+//  Created by Gaston Mazzeo on 21/11/2025.
+//
+
+protocol BondInteractor {
+    func fetchBondPositions(vault: Vault) async -> (active: [BondPosition], available: [BondNode])
+    func canUnbond() async -> Bool
+}
