+++ VultisigApp/VultisigApp/Features/Defi/DefiMain/View/DefiChainListView.swift
@@ -28,7 +28,7 @@ struct DefiChainListView: View {
             NavigationLink {
                 switch group.chain {
                 case .thorChain:
-                    DefiTHORChainMainScreen(vault: vault, group: group)
+                    DefiChainMainScreen(vault: vault, group: group)
                 default:
                     EmptyView()
                 }
